src/agents/requirements.txt
azure-identity
python-dotenv>=1.0.1
semantic-kernel==1.21.3
dapr>=1.14.0
fastapi
uvicorn
dapr-ext-fastapi>=1.14.0
azure-monitor-opentelemetry-exporter>=1.0.0b33